refactor: Removing hardcoded links to www.zensheet.com
On the previous version of the site, each html file
had on the menu an href reference to
www.zensheet.com/<subsite>. This has been replaced by
using a relative path as this was an issue when deploying
the site on a different domain (such as localhost or an AWS EBS one).

Some refactoring was made as well to have the HTML/JS/CSS files on a
'views' folder.

Now we are using the EJB template language.