feat: ZenSheet site redirects to authentication middleware
Now the ZenSheet site provides to links to signup or sign in with
the AWS Cognito Authentication Platform. The information that is
needed for this is fetched from environment variables that must
be defined on the host computer. Further documentation of this
procedure will be added to the documentation of this repository.